Updated Bool Extension
Swift added a Bool initializer from string that conflicted with the extension herein. I updated Vapor's to include an explicit fuzzy: parameter value, then updated the appropriate tests. We decided to keep ours, becuase the implemetation in swift is more limited and will not take values like 'y' or 'n'